+++ .gitignore
@@ -146,3 +146,4 @@ test.py
 wsgi.py
 
 attractionScriptv1.js
+test.*
